test(marketplace): change jest to vitest for better suport to esm
