+++ packages/marketplace/src/artifacts/contract-ids.json
@@ -1,0 +1,3 @@
+{
+  "marketplace": "0x3f1ec560123a1bbd929c03e4fde1cd54d4c3071a1efd00b427e4ad4b9ac730bb"
+}
