+++ packages/marketplace/src/sdk/index.ts
@@ -1,1 +1,2 @@
+export * from './client';
 export * from './marketplace';
