+++ packages/marketplace/src/sdk/marketplace.ts
@@ -1,11 +1,5 @@
 import { type Account, type BN, Provider } from 'fuels';
-import { GraphQLClient } from 'graphql-request';
 import { Marketplace } from '../artifacts';
-import {
-  type GQLAssetsQueryVariables,
-  type GQLOrdersQueryVariables,
-  getSdk,
-} from '../graphql/generated/operations';
 import { callAndWait, getContractId } from '../utils';
 
 /**
@@ -15,7 +9,7 @@ import { callAndWait, getContractId } from '../utils';
  * @property {BN} price - The price of the asset
  * @property {string} priceAsset - The asset to be used as the price
  */
-type Order = {
+export type Order = {
   itemAsset: string;
   itemAmount: BN;
   sellPrice: BN;
@@ -27,7 +21,7 @@ type Order = {
  * @property {BN} sellPrice - The price of the asset
  * @property {string} sellAsset - The asset to be used as the price
  */
-type UpdateOrder = {
+export type UpdateOrder = {
   sellPrice: BN;
   sellAsset: string;
 };
@@ -209,45 +203,5 @@ export class MarketplaceContract {
     return {
       transactionResult: response.transactionResult,
     };
-  }
-}
-
-/**
- * @description Marketplace client
- * @property {ReturnType<typeof getSdk>} sdk - The sdk
- */
-export class MarketplaceClient {
-  private readonly sdk: ReturnType<typeof getSdk>;
-  private readonly client: GraphQLClient;
-
-  /**
-   * @description Constructor
-   * @param {string} url - The url of the GraphQL API
-   */
-  constructor(url?: string) {
-    this.client = new GraphQLClient(url ?? process.env.GRAPHQL_API_URL!);
-    this.sdk = getSdk(this.client);
-  }
-
-  /**
-   * @description Get assets
-   * @param {GQLAssetsQueryVariables} variables - The variables
-   * @returns {Promise<Asset[]>} - The assets
-   */
-  async getAssets(variables?: GQLAssetsQueryVariables) {
-    const { data } = await this.sdk.assets(variables);
-
-    return data.Asset;
-  }
-
-  /**
-   * @description Get orders
-   * @param {GQLOrdersQueryVariables} variables - The variables
-   * @returns {Promise<Order[]>} - The orders
-   */
-  async getOrders(variables?: GQLOrdersQueryVariables) {
-    const { data } = await this.sdk.orders(variables);
-
-    return data.Order;
   }
 }
